fix: mv jwt to auth

+++ common/utils/jwt.go
@@ -8,7 +8,7 @@ import (
 	"github.com/golang-jwt/jwt/v4"
 )
 
-func GenerateToken(userID string, role string, issuer string, secret string, expire int64) (string, time.Time, error) {
+func GenerateToken(userID string, role string, issuer string, secret string, expire int) (string, time.Time, error) {
 	now := time.Now()
 	exp := now.Add(time.Second * time.Duration(expire))
 
@@ -85,7 +85,7 @@ func ParseToken(tokenString string, secret string) (string, string, error) {
 	return "", "", jwt.ErrTokenInvalidClaims
 }
 
-func RefreshToken(tokenString string, iss, secret string, expire int64) (string, time.Time, error) {
+func RefreshToken(tokenString string, iss, secret string, expire int) (string, time.Time, error) {
 	userID, role, err := ParseToken(tokenString, secret)
 	if err != nil {
 		return "", time.Now(), err
